Move logging facilities from -sdk to a new -backend-utils module
We intend to use the logger stuff in both -sdk and -engine, so move it
into -backend-utils, which will be a new common dependency for both -sdk
and -engine.

We are creating a new module here because we want to keep this stuff OUT
of -core and -client.  The reason for this is Winston only works in the
Node environment, and we would like -core and -client to be usable
directly in browsers (or at least, we don't want to preclude their use
in browsers, even if we aren't actively supporting it right now).