Make -engine use the logging from -backend-utils
This removes the direct dependency on Winston and replaces it with the
logging facilities in -backend-utils, which includes much nicer log
formats when running in debug mode.

We also refactor the auditing stuff a bit to more cleanly separate what
is auditing from what is regular logging--in particular, AuditLogger is
no longer a Logger, but uses a Logger internally.