Regenerate `package-lock.json`s
We are seeing intermittent issues in both local builds and CircleCI
where `npm install` doesn't work because of some missing file in
`node_modules/.staging`.  Not sure if this is related to some
inconsistency in the `package-lock.json` files, but I'm going to try
this and see if it works.

The updates to `package-lock.json` files caused some other type issues
due to upstream type definitions changing, so fix those up as well and
add some additional validation where it makes sense.  (This also fixes a
latent bug in validating offset/limit parameters on entity retrieval
operations, which was exposed by the additional validation.)