python SDK: Always provide argument/result when defining a procedure
If these properties are None/omitted, Papiea will not be able to show
the Swagger / API docs at all for this provider and would fail because
it was expecting to see an object where there was none.

NOTE: This really should have failed at provider registration time;
however, papiea-engine accepted the provider registration with these
items missing.