Makefile hacks to fix dependencies
The dependencies throughout the Makefile were improperly specified, so
nothing was ever getting rebuilt if it already existed.  For now, I have
made everything `.PHONY` to work around this, but we need to specify
proper dependencies on the non-PHONY targets.

Also, I have changed the `default` target to actually build the project,
rather than simply building the benchmarks.  This was rather surprising
(to say the least).

Also also, I broke up the giant `.PHONY` target into tagging individual
targets.  Having one large `.PHONY` is not best practice because it's
very hard to see whether an individual target has been tagged `.PHONY`
or not.