Key the watchlist on provider/kind/uuid, not just uuid
If there are different entities that share the same UUID (e.g. in
different providers/kinds), the watchlist could previously not
differentiate between the two, and would miss diffs on all but one of
the entities with the same UUID.

Fix this by changing the watchlist to be keyed on a combination of
provider, kind and UUID, not just UUID.

NOTE: This has been tested in vivo, but needs proper unit testing.