Watchlist and Python fixes (#446)
* python SDK: Always provide argument/result when defining a procedure

If these properties are None/omitted, Papiea will not be able to show
the Swagger / API docs at all for this provider and would fail because
it was expecting to see an object where there was none.

NOTE: This really should have failed at provider registration time;
however, papiea-engine accepted the provider registration with these
items missing.

* Makefile hacks to fix dependencies

The dependencies throughout the Makefile were improperly specified, so
nothing was ever getting rebuilt if it already existed.  For now, I have
made everything `.PHONY` to work around this, but we need to specify
proper dependencies on the non-PHONY targets.

Also, I have changed the `default` target to actually build the project,
rather than simply building the benchmarks.  This was rather surprising
(to say the least).

Also also, I broke up the giant `.PHONY` target into tagging individual
targets.  Having one large `.PHONY` is not best practice because it's
very hard to see whether an individual target has been tagged `.PHONY`
or not.

* Key the watchlist on provider/kind/uuid, not just uuid

If there are different entities that share the same UUID (e.g. in
different providers/kinds), the watchlist could previously not
differentiate between the two, and would miss diffs on all but one of
the entities with the same UUID.

Fix this by changing the watchlist to be keyed on a combination of
provider, kind and UUID, not just UUID.

NOTE: This has been tested in vivo, but needs proper unit testing.

* Comment out failing test case for now

* Comment out more tests

* fixed db test

* fixed multiple bugs and tests

* fix unhandled exception problem

* Tests shared UUID watchlist (#451)

* started adding tests

* fixed the first test

* finished 2nd test

* added but commented out test where a single provider with shared entity uuids are used. Waiting for clarifications

* fixed possible status update bug & changed uniqueness rules

* increase timeouts

* changed order of tests

* removed .PHONY from targets that are not .PHONY

* Revert "removed .PHONY from targets that are not .PHONY"

This reverts commit cfed5059a92f8484141ca60315a780328db2abc4.

Co-authored-by: Shlomi Vaknin <shlomi.vaknin@nutanix.com>
Co-authored-by: igorkorotach <igorkorotach@gmail.com>